requirements/base.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements/base.txt requirements/base.in
#
-e git+https://github.com/elyak123/django-pwa.git@v1.0.11#egg=django-pwa  # via -r requirements/base.in
asn1crypto==1.4.0         # via oscrypto
babel==2.9.0              # via django-phonenumber-field
beautifulsoup4==4.9.3     # via django-bootstrap4
certifi==2020.12.5        # via requests
cffi==1.14.4              # via validate-aws-sns-message
chardet==3.0.4            # via requests
defusedxml==0.6.0         # via python3-openid
dj-rest-auth==1.0.1       # via -r requirements/base.in
django-admin-interface==0.13.2  # via -r requirements/base.in
django-allauth==0.41.0    # via -r requirements/base.in
django-appconf==1.0.4     # via django-compressor
django-bootstrap4==1.1.1  # via -r requirements/base.in
django-colorfield==0.3.2  # via django-admin-interface
django-dynamic-model-validation==0.1.13  # via -r requirements/base.in
django-environ==0.4.5     # via -r requirements/base.in
django-flat-responsive==2.0  # via django-admin-interface
django-flat-theme==1.1.4  # via django-admin-interface
django-floppyforms==1.9.0  # via -r requirements/base.in
django-hosts==4.0         # via -r requirements/base.in
django-phonenumber-field==4.0.0  # via -r requirements/base.in
django==2.2.10            # via -r requirements/base.in, dj-rest-auth, django-allauth, django-appconf, django-phonenumber-field, django-pwa, djangorestframework-simplejwt
django_compressor==2.4    # via -r requirements/base.in
djangorestframework-gis==0.15  # via -r requirements/base.in
djangorestframework-simplejwt==4.4.0  # via -r requirements/base.in
djangorestframework==3.10.3  # via -r requirements/base.in, dj-rest-auth, djangorestframework-gis, djangorestframework-simplejwt
future==0.17.1            # via django-dynamic-model-validation
idna==2.10                # via requests
oauthlib==3.1.0           # via requests-oauthlib
openrouteservice==2.3.0   # via -r requirements/base.in
oscrypto==1.2.1           # via validate-aws-sns-message
phonenumbers==8.12.1      # via -r requirements/base.in
psycopg2==2.8.6           # via -r requirements/base.in
pycparser==2.20           # via cffi
pyjwt==1.7.1              # via djangorestframework-simplejwt
python3-openid==3.2.0     # via django-allauth
pytz==2020.4              # via babel, django
rcssmin==1.0.6            # via django-compressor
requests-oauthlib==1.3.0  # via django-allauth
requests==2.23.0          # via -r requirements/base.in, django-allauth, openrouteservice, requests-oauthlib
rjsmin==1.1.0             # via django-compressor
six==1.14.0               # via django-admin-interface, django-compressor, django-dynamic-model-validation, validate-aws-sns-message
soupsieve==2.1            # via beautifulsoup4
sqlparse==0.4.1           # via django
unidecode==1.1.1          # via -r requirements/base.in
urllib3==1.25.11          # via requests
validate_aws_sns_message==0.9.0  # via -r requirements/base.in
xkcdpass==1.17.3          # via -r requirements/base.in
